+++ src/serde/mod.rs
@@ -8,6 +8,7 @@ mod serialize;
 mod str;
 
 #[cfg(feature = "metadata")]
+#[cfg(feature = "serde")]
 mod metadata;
 
 pub use de::from_dynamic;